+++ lib/components/constants/text_string.dart
@@ -1,0 +1,89 @@
+/// This class contains all the App Text in String formats.
+class TTexts {
+  // -- Admin Default Credentials
+  static const String adminEmail = 'support@motocode.com.ph';
+  static const String adminPassword = 'Admin@123';
+
+  // -- GLOBAL Texts
+  static const String and = 'and';
+  static const String apply = 'Apply';
+  static const String skip = 'Skip';
+  static const String done = 'Done';
+  static const String update = 'update';
+  static const String submit = 'Submit';
+  static const String appName = 'Motocode Portal';
+  static const String tContinue = 'Continue';
+
+  // -- Authentication Forms
+  static const String firstName = 'First Name';
+  static const String lastName = 'Last Name';
+  static const String email = 'E-Mail';
+  static const String password = 'Password';
+  static const String newPassword = 'New Password';
+  static const String confirmPassword = 'Confirm Password';
+  static const String username = 'Username';
+  static const String phoneNo = 'Phone Number';
+  static const String rememberMe = 'Remember Me';
+  static const String forgetPassword = 'Forget Password?';
+  static const String processingRequest = 'Processing your request ...';
+  static const String underDevelopment =
+      'This page is under development. Please check back soon.';
+  static const String signIn = 'Sign In';
+  static const String createAccount = 'Create Account';
+  static const String orSignInWith = 'or sign in with';
+  static const String orSignUpWith = 'or sign up with';
+  static const String iAgreeTo = 'I agree to';
+  static const String privacyPolicy = 'Privacy Policy';
+  static const String termsOfUse = 'Terms of use';
+  static const String verificationCode = 'verificationCode';
+  static const String resendEmail = 'Resend Email';
+  static const String toLoginPage = 'Go back to login';
+  static const String resendEmailIn = 'Resend email in';
+
+  // -- Authentication Headings
+  static const String loginTitle = 'Welcome back,';
+  static const String loginsubTitle = 'Motorcycle Tracking System';
+  static const String resetPasswordTitle = 'Reset Your Password,';
+  static const String resetPasswordSubTitle =
+      'Enter a new password to regain access to your account. Make sure it’s strong and secure!';
+  static const String loginSubTitle =
+      'Where motorcycle management and tracking happen.';
+  static const String signupTitle = 'Let’s create your account';
+  static const String forgetPasswordTitle = 'Forget password';
+  static const String forgetPasswordSubTitle =
+      'Don’t worry sometimes people can forget too, enter your email and we will send you a password reset link.';
+  static const String changeYourPasswordTitle = 'Password Reset Email Sent';
+  static const String changeYourPasswordSubTitle =
+      "Your Account Security is Our Priority! We've Sent You a Secure Link to Safely Change Your Password and Keep Your Account Protected.";
+  static const String confirmEmail = 'Verify your email address!';
+  static const String confirmEmailSubTitle =
+      'Congratulations! Your Account Awaits: Verify Your Email to Start Shopping and Experience a World of Unrivaled Deals and Personalized Offers.';
+  static const String emailNotReceivedMessage =
+      'Didn’t get the email? Check your junk/spam or resend it.';
+  static const String yourAccountCreatedTitle =
+      'Your account successfully created!';
+
+  //Upload VINS
+  static const String uploadsubTitle = 'Upload an excel file.';
+  static const String uploadTitle = 'Upload File';
+  static const String uploadPreviewTitle = 'File Content Preview';
+  static const String uploadPreviewsubTitle =
+      'You can now verify the file records once the data appears in this area';
+  static const String uploadContentTitle =
+      'Files you uploaded will show up here...';
+
+  // -- Create Account
+  static const String branchSubTitle =
+      'Select branch to manage by this account';
+  static const String selectedBranchSubTitle =
+      'Branches managed by this account.';
+  static const String emptySelectedBranchText = 'Please assign a branch.';
+  static const String errorText =
+      'Oops! Something went wrong while loading resources. Please refresh the page.';
+
+  // -- Infos
+  static const String infoMessage = 'To update your status, goto edit account.';
+
+  static const String infoAccountMessage =
+      'To create new account, goto create account page.';
+}